Names are a great help with Liskof Substition
$report = new TabloidReport() makes sense however the Report classes name is now too vague.
Make it better.

There are two more classes that need renaming... do that.

Now that we have renamed the classes have a think about what is now missing.

What would Barbara do?

Also, check that all your derived types (on the right of the =) have
more specific names than the names on the left.