Interface decomposition, a powerful tool
If you look you will see DataAccess implements an Interface, but does
not implement most of the functions as it does not need to. this is a
problem as another class may call these functions and break our program.

There may be other classes using IDataAccess we cannot touch it.
Instead create 3 new interfaces called IGetData, IPersistData and IDeleteData
and split out the functions.

Now we can get rid of the problem functions in DataAccess.